ParamedicSauceLabs: all Sauce Labs only methods extracted from paramedic.js
