+++ lib/paramedic.js
@@ -25,17 +25,12 @@ var Server          = require('./LocalServer');
 var path            = require('path');
 var Q               = require('q');
 var fs              = require('fs');
-var wd              = require('wd');
-var SauceLabs       = require('saucelabs');
-var randomstring    = require('randomstring');
-var sauceConnectLauncher    = require('sauce-connect-launcher');
 
 var logger          = require('./utils').logger;
 var util            = require('./utils').utilities;
 var Reporters       = require('./Reporters');
 var ParamedicKill   = require('./ParamedicKill');
 var AppiumRunner    = require('./appium/AppiumRunner');
-var appPatcher      = require('./appium/helpers/appPatcher');
 var ParamedicLogCollector   = require('./ParamedicLogCollector');
 var ParamediciOSPermissions = require('./ParamediciOSPermissions');
 var ParamedicTargetChooser  = require('./ParamedicTargetChooser');
@@ -49,6 +44,7 @@ require('./appium/helpers/wdHelper');
 // If device has not connected within this interval the tests are stopped.
 var INITIAL_CONNECTION_TIMEOUT = 540000; // 9mins
 
+Q.longStackSupport = true;
 
 function ParamedicRunner(config, _callback) {
     this.tempFolder = null;
@@ -57,6 +53,8 @@ function ParamedicRunner(config, _callback) {
     this.targetObj = undefined;
 
     exec.setVerboseLevel(config.isVerbose());
+
+    this.paramedicSauceLabs = null;
 }
 
 ParamedicRunner.prototype.run = function () {
@@ -96,6 +94,11 @@ ParamedicRunner.prototype.run = function () {
         return self.runTests();
     })
     .timeout(self.config.getTimeout(), 'Timed out after waiting for ' + self.config.getTimeout() + ' ms.')
+    .catch(function (error) {
+        logger.error(error);
+        console.log(error.stack);
+        throw new Error(error);
+    })
     .fin(function (result) {
         isTestPassed = result;
         logger.normal('Completed tests at ' + (new Date()).toLocaleTimeString());
@@ -120,7 +123,10 @@ ParamedicRunner.prototype.run = function () {
 };
 
 ParamedicRunner.prototype.checkConfig = function () {
-    this.checkSauceRequirements();
+    if (this.config.shouldUseSauce()) {
+        this.paramedicSauceLabs = new ParamedicSauceLabs();
+        this.paramedicSauceLabs.checkSauceRequirements();
+    }
     if (!this.config.runMainTests() && !this.config.runAppiumTests()) {
         throw new Error('No tests to run: both --skipAppiumTests and --skipMainTests are used');
     }
@@ -476,74 +482,6 @@ ParamedicRunner.prototype.cleanUpProject = function () {
     }
 };
 
-ParamedicRunner.prototype.checkSauceRequirements = function () {
-    if (this.config.shouldUseSauce()) {
-        var platformId = this.config.getPlatformId();
-        if (platformId !== util.ANDROID && platformId !== util.IOS && platformId !== util.BROWSER) {
-            logger.warn('Saucelabs only supports Android and iOS (and browser), falling back to testing locally.');
-            this.config.setShouldUseSauce(false);
-        } else if (!this.config.getSauceKey()) {
-            throw new Error('Saucelabs key not set. Please set it via environmental variable ' +
-                util.SAUCE_KEY_ENV_VAR + ' or pass it with the --sauceKey parameter.');
-        } else if (!this.config.getSauceUser()) {
-            throw new Error('Saucelabs user not set. Please set it via environmental variable ' +
-                util.SAUCE_USER_ENV_VAR + ' or pass it with the --sauceUser parameter.');
-        } else if (!this.shouldWaitForTestResult()) {
-            // don't throw, just silently disable Sauce
-            this.config.setShouldUseSauce(false);
-        }
-    }
-};
-
-ParamedicRunner.prototype.packageApp = function () {
-    var self = this;
-    switch (this.config.getPlatformId()) {
-        case util.IOS: {
-            return Q.Promise(function (resolve, reject) {
-                var zipCommand = 'zip -r ' + self.getPackageName() + ' ' + self.getBinaryName();
-                shell.pushd(self.getPackageFolder());
-                shell.rm('-rf', self.getPackageName());
-                console.log('Running command: ' + zipCommand + ' in dir: ' + shell.pwd());
-                exec(zipCommand, function (code, stdout, stderr) {
-                    shell.popd();
-                    if (code) {
-                        reject('zip command returned with error code ' + code);
-                    } else {
-                        resolve();
-                    }
-                });
-            });
-        }
-        case util.ANDROID:
-            break; // don't need to zip the app for Android
-        case util.BROWSER:
-            break; // don't need to bundle the app on Browser platform at all
-        default:
-            throw new Error('Don\'t know how to package the app for platform: ' + this.config.getPlatformId());
-    }
-    return Q.resolve();
-};
-
-ParamedicRunner.prototype.uploadApp = function () {
-    logger.normal('cordova-paramedic: uploading ' + this.getAppName() + ' to Sauce Storage');
-
-    var sauceUser = this.config.getSauceUser();
-    var key       = this.config.getSauceKey();
-
-    var uploadURI     = encodeURI('https://saucelabs.com/rest/v1/storage/' + sauceUser + '/' + this.getAppName() + '?overwrite=true');
-    var filePath      = this.getPackagedPath();
-    var uploadCommand =
-        'curl -u ' + sauceUser + ':' + key +
-        ' -X POST -H "Content-Type: application/octet-stream" ' +
-        uploadURI + ' --data-binary "@' + filePath + '"';
-
-    return execPromise(uploadCommand);
-};
-
-ParamedicRunner.prototype.getPackagedPath = function () {
-    return path.join(this.getPackageFolder(), this.getPackageName());
-};
-
 ParamedicRunner.prototype.killEmulatorProcess = function () {
     if(this.config.shouldCleanUpAfterRun()){
         logger.info('cordova-paramedic: Killing the emulator process.');
@@ -564,421 +502,6 @@ ParamedicRunner.prototype.uninstallApp = function () {
     logger.info('Uninstalling the app.');
     var paramedicAppUninstall = new ParamedicAppUninstall(this.tempFolder.name, this.config.getPlatformId());
     return paramedicAppUninstall.uninstallApp(this.targetObj,util.PARAMEDIC_DEFAULT_APP_NAME);
-};
-
-ParamedicRunner.prototype.getPackageFolder = function () {
-    var packageDirs = this.getPackageFolders();
-    var foundDir = null;
-    packageDirs.forEach (function (dir) {
-        if (fs.existsSync(dir) && fs.statSync(dir).isDirectory()) {
-            foundDir = dir;
-            return;
-        }
-    });
-    if (foundDir != null) {
-        return foundDir;
-    }
-    throw new Error ('Couldn\'t locate a built app directory. Looked here: ' + packageDirs);
-};
-
-ParamedicRunner.prototype.getPackageFolders = function () {
-    var packageFolders;
-    switch (this.config.getPlatformId()) {
-        case util.ANDROID:
-            packageFolders =  [ path.join(this.tempFolder.name, 'platforms/android/app/build/outputs/apk/debug'),
-                                path.join(this.tempFolder.name, 'platforms/android/build/outputs/apk') ];
-            break;
-        case util.IOS:
-            packageFolders = [ path.join(this.tempFolder.name, 'platforms/ios/build/emulator') ];
-            break;
-        default:
-            throw new Error('Don\t know where the package foler is for platform: ' + this.config.getPlatformId());
-    }
-    return packageFolders;
-};
-
-ParamedicRunner.prototype.getPackageName = function () {
-    var packageName;
-    switch (this.config.getPlatformId()) {
-        case util.IOS:
-            packageName = 'HelloCordova.zip';
-            break;
-        case util.ANDROID:
-            packageName = this.getBinaryName();
-            break;
-        default:
-            throw new Error('Don\'t know what the package name is for platform: ' + this.config.getPlatformId());
-    }
-    return packageName;
-};
-
-ParamedicRunner.prototype.getBinaryName = function () {
-    var binaryName;
-    switch (this.config.getPlatformId()) {
-        case util.ANDROID:
-            shell.pushd(this.getPackageFolder());
-            var apks = shell.ls('*debug.apk');
-            if (apks.length > 0) {
-                binaryName = apks.reduce(function (previous, current) {
-                    // if there is any apk for x86, take it
-                    if (current.indexOf('x86') >= 0) {
-                        return current;
-                    }
-                    // if not, just take the first one
-                    return previous;
-                });
-            } else {
-                throw new Error('Couldn\'t locate built apk');
-            }
-            shell.popd();
-            break;
-        case util.IOS:
-            binaryName = 'HelloCordova.app';
-            break;
-        default:
-            throw new Error('Don\'t know the binary name for platform: ' + this.config.getPlatformId());
-    }
-    return binaryName;
-};
-
-// Returns a name of the file at the SauceLabs storage
-ParamedicRunner.prototype.getAppName = function () {
-    if (this.appName) {
-        return this.appName;
-    }
-    var appName = randomstring.generate();
-    switch (this.config.getPlatformId()) {
-        case util.ANDROID:
-            appName += '.apk';
-            break;
-        case util.IOS:
-            appName += '.zip';
-            break;
-        default:
-            throw new Error('Don\'t know the app name for platform: ' + this.config.getPlatformId());
-    }
-    this.appName = appName;
-    return appName;
-};
-
-ParamedicRunner.prototype.displaySauceDetails = function (buildName) {
-    if (!this.config.shouldUseSauce()) {
-        return Q();
-    }
-    if (!buildName) {
-        buildName = this.config.getBuildName();
-    }
-
-    var self = this;
-    var d = Q.defer();
-
-    logger.normal('Getting saucelabs jobs details...\n');
-
-    var sauce = new SauceLabs({
-        username: self.config.getSauceUser(),
-        password: self.config.getSauceKey()
-    });
-
-    sauce.getJobs(function (err, jobs) {
-        var found = false;
-        for (var job in jobs) {
-            if (jobs.hasOwnProperty(job) && jobs[job].name && jobs[job].name.indexOf(buildName) === 0) {
-                var jobUrl = 'https://saucelabs.com/beta/tests/' + jobs[job].id;
-                logger.normal('============================================================================================');
-                logger.normal('Job name: ' + jobs[job].name);
-                logger.normal('Job ID: ' + jobs[job].id);
-                logger.normal('Job URL: ' + jobUrl);
-                logger.normal('Video: ' + jobs[job].video_url);
-                logger.normal('Appium logs: ' + jobs[job].log_url);
-                if (self.config.getPlatformId() === util.ANDROID) {
-                    logger.normal('Logcat logs: ' + 'https://saucelabs.com/jobs/' + jobs[job].id + '/logcat.log');
-                }
-                logger.normal('============================================================================================');
-                logger.normal('');
-                found = true;
-            }
-        }
-
-        if (!found) {
-            logger.warn('Can not find saucelabs job. Logs and video will be unavailable.');
-        }
-        d.resolve();
-    });
-    return d.promise;
-};
-
-ParamedicRunner.prototype.getSauceCaps = function () {
-    this.sauceBuildName = this.sauceBuildName || this.config.getBuildName();
-    var caps = {
-        name: this.sauceBuildName,
-        idleTimeout: '100', // in seconds
-        maxDuration: util.SAUCE_MAX_DURATION,
-        tunnelIdentifier: this.config.getSauceTunnelId(),
-    };
-
-    switch(this.config.getPlatformId()) {
-        case util.ANDROID:
-            caps.platformName = 'Android';
-            caps.appPackage = 'io.cordova.hellocordova';
-            caps.appActivity = 'io.cordova.hellocordova.MainActivity';
-            caps.app = 'sauce-storage:' + this.getAppName();
-            caps.deviceType = 'phone';
-            caps.deviceOrientation = 'portrait';
-            caps.appiumVersion = this.config.getSauceAppiumVersion();
-            caps.deviceName = this.config.getSauceDeviceName();
-            caps.platformVersion = this.config.getSaucePlatformVersion();
-            break;
-        case util.IOS:
-            caps.platformName = 'iOS';
-            caps.autoAcceptAlerts = true;
-            caps.waitForAppScript = 'true;';
-            caps.app = 'sauce-storage:' + this.getAppName();
-            caps.deviceType = 'phone';
-            caps.deviceOrientation = 'portrait';
-            caps.appiumVersion = this.config.getSauceAppiumVersion();
-            caps.deviceName = this.config.getSauceDeviceName();
-            caps.platformVersion = this.config.getSaucePlatformVersion();
-            break;
-        case util.BROWSER:
-            caps.browserName = this.config.getSauceDeviceName() || 'chrome';
-            caps.version = this.config.getSaucePlatformVersion() || '45.0';
-            caps.platform = caps.browserName.indexOf('Edge') > 0 ? 'Windows 10' : 'macOS 10.13';
-            // setting from env.var here and not in the config
-            // because for any other platform we don't need to put the sauce connect up 
-            // unless the tunnel id is explicitly passed (means that user wants it anyway)
-            if (!caps.tunnelIdentifier && process.env[util.SAUCE_TUNNEL_ID_ENV_VAR]) {
-                caps.tunnelIdentifier = process.env[util.SAUCE_TUNNEL_ID_ENV_VAR];
-            } else if (!caps.tunnelIdentifier) {
-                throw new Error('Testing browser platform on Sauce Labs requires Sauce Connect tunnel. Please specify tunnel identifier via --sauceTunnelId');
-            }
-            break;
-        default:
-            throw new Error('Don\'t know the Sauce caps for platform: ' + this.config.getPlatformId());
-    }
-    return caps;
-};
-
-ParamedicRunner.prototype.connectWebdriver = function () {
-    var user = this.config.getSauceUser();
-    var key = this.config.getSauceKey();
-    var caps = this.getSauceCaps();
-
-    logger.normal('cordova-paramedic: connecting webdriver');
-    var spamDots = setInterval(function () {
-        process.stdout.write('.');
-    }, 1000);
-
-    wd.configureHttp({
-        timeout: util.WD_TIMEOUT,
-        retryDelay: util.WD_RETRY_DELAY,
-        retries: util.WD_RETRIES
-    });
-
-    var driver = wd.promiseChainRemote(util.SAUCE_HOST, util.SAUCE_PORT, user, key);
-    return driver
-        .init(caps)
-        .then(function () {
-            clearInterval(spamDots);
-            process.stdout.write('\n');
-        }, function (error) {
-            clearInterval(spamDots);
-            process.stdout.write('\n');
-            throw(error);
-        });
-};
-
-ParamedicRunner.prototype.connectSauceConnect = function () {
-    var self = this;
-    var isBrowser = self.config.getPlatformId() === util.BROWSER;
-
-    // on platforms other than browser, only run sauce connect if user explicitly asks for it
-    if (!isBrowser && !self.config.getSauceTunnelId()) {
-        return Q();
-    }
-    // on browser, run sauce connect in any case
-    if (isBrowser && !self.config.getSauceTunnelId()) {
-        self.config.setSauceTunnelId(process.env[util.SAUCE_TUNNEL_ID_ENV_VAR] || self.config.getBuildName());
-    }
-
-    return Q.Promise(function (resolve, reject) {
-        logger.info('cordova-paramedic: Starting Sauce Connect...');
-        sauceConnectLauncher({
-            username: self.config.getSauceUser(),
-            accessKey: self.config.getSauceKey(),
-            tunnelIdentifier: self.config.getSauceTunnelId(),
-            connectRetries: util.SAUCE_CONNECT_CONNECTION_RETRIES,
-            connectRetryTimeout: util.SAUCE_CONNECT_CONNECTION_TIMEOUT,
-            downloadRetries: util.SAUCE_CONNECT_DOWNLOAD_RETRIES,
-            downloadRetryTimeout: util.SAUCE_CONNECT_DOWNLOAD_TIMEOUT,
-        }, function (err, sauceConnectProcess) {
-            if (err) {
-                reject(err);
-            }
-            self.sauceConnectProcess = sauceConnectProcess;
-            logger.info('cordova-paramedic: Sauce Connect ready');
-            resolve();
-        });
-    });
-};
-
-ParamedicRunner.prototype.runSauceTests = function () {
-    var self = this;
-    var isTestPassed = false;
-    var pollForResults;
-    var driver;
-    var runProcess = null;
-
-    if (!self.config.runMainTests()) {
-        logger.normal('Skipping main tests...');
-        return Q(util.TEST_PASSED);
-    }
-
-    logger.info('cordova-paramedic: running tests with sauce');
-
-    return Q().then(function () {
-        // Build + "Upload" app
-        if (self.config.getPlatformId() === util.BROWSER) {
-            // for browser, we need to serve the app for Sauce Connect
-            // we do it by just running "cordova run" and ignoring the chrome instance that pops up
-            return Q()
-                .then(function() {
-                    appPatcher.addCspSource(self.tempFolder.name, 'connect-src', 'http://*');
-                    appPatcher.permitAccess(self.tempFolder.name, '*');
-                    return self.getCommandForStartingTests();
-                })
-                .then(function (command) {
-                    console.log('$ ' + command);
-                    runProcess = cp.exec(command, function onExit() {
-                        // a precaution not to try to kill some other process
-                        runProcess = null;
-                    });
-                });
-        } else {
-            return self.buildApp()
-                .then(self.packageApp.bind(self))
-                .then(self.uploadApp.bind(self));
-        }
-    })
-    .then(function () {
-        return self.connectSauceConnect();
-    })
-    .then(function () {
-        driver = self.connectWebdriver();
-        if (self.config.getPlatformId() === util.BROWSER) {
-            return driver.get('http://localhost:8000/cdvtests/index.html');
-        }
-        return driver;
-    })
-    .then(function () {
-        if (self.config.getUseTunnel() || self.config.getPlatformId() === util.BROWSER) {
-            return driver;
-        }
-        return driver
-        .getWebviewContext()
-        .then(function (webview) {
-            return driver.context(webview);
-        });
-    })
-    .then(function () {
-        var isWkWebview = false;
-        var plugins = self.config.getPlugins();
-        for (var plugin in plugins) {
-            if (plugins[plugin].indexOf('wkwebview') >= 0) {
-                isWkWebview = true;
-            }
-        }
-        if (isWkWebview) {
-            logger.normal('cordova-paramedic: navigating to a test page');
-            return driver
-                .sleep(1000)
-                .elementByXPath('//*[text() = "Auto Tests"]')
-                .click();
-        }
-        return driver;
-    })
-    .then(function () {
-        logger.normal('cordova-paramedic: connecting to app');
-
-        var platform = self.config.getPlatformId();
-        var plugins = self.config.getPlugins();
-
-        var skipBuster = false;
-        // skip permission buster for splashscreen and inappbrowser plugins
-        // it hangs the test run on Android 7 for some reason
-        for (var i = 0; i < plugins.length; i++) {
-            if (plugins[i].indexOf('cordova-plugin-splashscreen') >= 0 || plugins[i].indexOf('cordova-plugin-inappbrowser') >= 0) {
-                skipBuster = true;
-            }
-        }
-        // always skip buster for browser platform
-        if (platform === util.BROWSER) {
-            skipBuster = true;
-        }
-
-        if (!self.config.getUseTunnel()) {
-            var polling = false;
-            pollForResults = setInterval(function () {
-                if (!polling) {
-                    polling = true;
-                    driver.pollForEvents(platform, skipBuster)
-                    .then(function (events) {
-                        for (var i = 0; i < events.length; i++) {
-                            self.server.emit(events[i].eventName, events[i].eventObject);
-                        }
-                        polling = false;
-                    })
-                    .fail(function (error) {
-                        logger.warn('appium: ' + error);
-                        polling = false;
-                    });
-                }
-            }, 2500);
-        }
-
-        return self.waitForTests();
-    })
-    .then(function (result) {
-        logger.normal('cordova-paramedic: Tests finished');
-        isTestPassed = result;
-    }, function (error) {
-        logger.normal('cordova-paramedic: Tests failed to complete; ending appium session. The error is:\n' + error.stack);
-    })
-    .fin(function () {
-        if (pollForResults) {
-            clearInterval(pollForResults);
-        }
-        if (driver && typeof driver.quit === 'function') {
-            return driver.quit();
-        }
-    })
-    .fin(function () {
-        if (self.config.getPlatformId() === util.BROWSER && !self.browserPatched) {
-            // we need to kill chrome
-            self.killEmulatorProcess();
-        }
-        if (runProcess) {
-            // as well as we need to kill the spawned node process serving our app
-            return Q.Promise(function (resolve, reject) {
-                util.killProcess(runProcess.pid, function () {
-                    resolve();
-                });
-            });
-        }
-    })
-    .fin(function () {
-        if (self.sauceConnectProcess) {
-            logger.info('cordova-paramedic: Closing Sauce Connect process...');
-            return Q.Promise(function (resolve, reject) {
-                self.sauceConnectProcess.close(function () {
-                    logger.info('cordova-paramedic: Successfully closed Sauce Connect process');
-                    resolve();
-                });
-            });
-        }
-    })
-    .then(function () {
-        return isTestPassed;
-    });
 };
 
 var storedCWD = null;
